--- src/components/explore-page/index.ts
@@ -1,1 +1,0 @@
-export * from './pagination';
