+++ src/components/shared/index.ts
@@ -2,6 +2,7 @@ export * from './button';
 export * from './separator';
 export * from './search-input';
 export * from './guide';
+export * from './pagination';
 
 // eslint no-cycle 오류 때문에 export 안함
 // export * from './layout-with-header';
