+++ src/containers/pages/explore/explore.container.tsx
@@ -1,9 +1,13 @@
 import { useReducer } from 'react';
 import { getVideos } from '@/apis/videos';
-import { Pagination } from '@/components/explore-page';
-import { usePagination } from '@/components/explore-page/pagination/hooks';
-import { SearchInput, Separator, useSearchInput } from '@/components/shared';
+import {
+  Pagination,
+  SearchInput,
+  Separator,
+  useSearchInput,
+} from '@/components/shared';
 import { VideoList } from '@/components/shared/video-list';
+import { usePagination } from '@/components/shared/pagination/hooks';
 import { categoryFilterItems, levelFilterItems } from './constants';
 import { useDetectCategoryFromParam, useGetVideos } from './hooks';
 import {
